Bugfix/#17436 fixing segment identification for setup (#31)
Segmentation made/delete correctly by @ljblancoredborder
