+++ resources/lib/wiz_conf.rb
@@ -1,0 +1,25 @@
+class WizConf
+
+    # Read propierties from sysfs for a network devices
+    def netdev_property(devname)
+        netdev = {}
+        IO.popen("udevadm info -q property -p /sys/class/net/#{devname} 2>/dev/null").each do |line|
+            unless line.match(/^(?<key>[^=]*)=(?<value>.*)$/).nil?
+                netdev[line.match(/^(?<key>[^=]*)=(?<value>.*)$/)[:key]] = line.match(/^(?<key>[^=]*)=(?<value>.*)$/)[:value]
+            end
+        end
+        if File.exist?"/sys/class/net/#{devname}/address"
+            f = File.new("/sys/class/net/#{devname}/address",'r')
+            netdev["MAC"] = f.gets.chomp
+            f.close
+        end
+        if File.exist?"/sys/class/net/#{devname}/operstate"
+            f = File.new("/sys/class/net/#{devname}/operstate",'r')
+            netdev["STATUS"] = f.gets.chomp
+            f.close
+        end
+
+        netdev
+    end
+
+end
